Rename UnixSocketAddr::unspecified() to new_unspecified()
Makes it clear that it returns a copy instead of being a method like
C++'s std::vector::empty().
new makes it sound more expensivee than it is,
but then again String::new() is also a no-op.

Should maybe be replaced with a const or static in the future,
(requires mem::zeroed() to become const fn),
but it might also be that it is quicker to create than copy :)