Fix getting peer credentials on FreeBSD
Was using the wrong socket level,
which resulted in issuing a getsockopt(SO_DEBUG) instead!
Since SO_DEBUG was not set, zero was re-written to the
start of xucred, and the rest was left untouched.

Fixing that made tests start failing,
because they weren't expecting non-root group IDs
or was too damn OS-specific.