Check for that some OSes might return the sent fd as is without cloning it
This means that closing the received fd will also close the original!
It makes actually testing fd-passing code nedlessy complex
and I don't understand why anyone writing a robust applicationwould want it.

I originally discovered this issue on DragonFly BSD
when I started working on adding support for that OS in March,
and it might also be the cause of a strange macOS CI failure:
https://travis-ci.com/github/tormol/uds/jobs/384392289