Adapt timeout methods and tests to other OSes and configurations
There was a reasen std and mio doesn't enable setting
or getting timeouts for listeners:
Only on Linux does accept() consider it.
Make sure setting it never silently succeeds on
operating systems where it is known not to work.

Timeouts for unix domain sockets is completely broken
on OmniOS (and I assume for all Illumos variants).

Also match std's Duration-to-timeval conversion and error message.